Add support for table controls in StoryView
This solution might be a bit convoluted, and could use refactoring.

It fixes a peculiar issue, where after a dropdown option is selected, because `Sift.Dictionary.merge` was called on controls; the control would change from type `array` to `string`.

This is due to a bigger issue, in which we erroneously assume the `controls` variable to be both the current settings and the potential settings.