Fix Searchbar filtering (#231)
# Problem

The current implementation of the filtering functionality in our
Storybook tree view has a limitation. It only checks if the name of the
current component node matches the filter text. This means that even if
a story matches the filter, if any of its parent folders does not match
the filter, the entire subtree gets skipped during rendering.

# Solution

The proposed solution modifies this behavior for non-story nodes (e.g.,
storybooks, folders). Now, instead of checking if the node's name
matches the filter, we check if the node has at least one descendant
story that matches the filter. This ensures that folders containing
matching stories are not erroneously hidden due to the folder name not
matching the filter.

For story nodes, the behavior remains the same as before: the story is
displayed if its name matches the filter.

Fix #161 

# Checklist

- [ ] Ran `./bin/test.sh` locally before merging