Patch for Search Filtering Feature (#237)
# Problem

This PR is a continuation of the work done in PR #231. It aims to
address the incomplete fix I implemented in my previous PR.

The problem addressed by this PR pertains to the handling of 'story'
nodes in our component tree. In the current implementation, 'story'
nodes, which inherently do not have any children, are incorrectly
subjected to a descendants check. This results in the `isEmpty` flag
always being set to true for 'Story' nodes.

# Solution

The main changes in this PR include:

- Refactoring the search filtering logic into a separate module.
- Adding unit tests

While the changes introduced in this PR could have been implemented
within the existing component, I chose to refactor them into a separate
module, so it's easier to test.

# Checklist

- [x] Ran `./bin/test.sh` locally
- [x] Ran `./bin/analyze.sh` locally

---------

Co-authored-by: Marin Minnerly <me@vocksel.com>