Update documentation in preparation for next release (#276)
# Problem

Our docs need an overhaul to reflect the upcoming release

# Solution

This is a big PR so I'll try to break things down.

First of all here's what our new sidebar looks like:
![Screenshot 2024-11-16 at 9 07
23 AM](https://github.com/user-attachments/assets/4560cd03-89b0-415e-a8a2-6f1d120ac51c)

And here's a summary of the changes:
1. Introduced `code-samples` package to the root of the repo. 
1. This is a package containing various examples for use in our
documentation. The kicker is these are run through analysis and linting
just like any of our other Luau code, so our code samples in docs are
easy to validate for correctness. It can also be synced to Studio and
all of the included stories can be viewed in Flipbook
3. Docs on writing stories have been rewritten
4. Updated docs on our Story/Storybook formats to reflect changes from
Storyteller
5. Docs have been added for writing React, Fusion, and Roact stories
along with tabs to switch to Storyteller variants
6. Updated docs for contributors
7. Minor change to our build script which now makes sure Flipbook's
storybook does not leak through to prod builds

# Checklist

- [x] Ran `lune run test` locally before merging