Inject build information when compiling (#287)
# Problem

It can be useful to know the version, channel, and commit hash for a
given build of Flipbook, especially to help when debugging. We only
include the version right now, the other two would be useful

# Solution

Created a new BuildInfo component which gets mounted on AboutView. It
leverages some new globals that get injected during the darklua build
step.

I've also removed the inclusion of `wally.toml` in the Flipbook plugin
build since the only info we were using was the version, and that's now
supplied via `_G.BUILD_VERSION`

# Checklist

- [x] Ran `lune run test` locally before merging