New tree view (#295)
# Problem

Our tree view is one of the oldest parts of the plugin and needs some
upgrades before we can solve issues like #275, #282, and #291

# Solution

I've created a new tree view from scratch that should support all the
issues mentioned above.

The tree view itself is fairly generic with the intent of supporting
future cases that require manipulating the structure beyond just
mimicking the DataModel. And due to it being general it should make a
good candidate to be packaged up later.

The StorybookTreeView component sits on top of the tree view to build
out the tree of storybooks and stories. I've also introduced a new user
setting to remember the last opened story. Now when reopening Flipbook
the last story is restored

Resolves #275

# Checklist

- [ ] Resolve issues with search
- [ ] Ran `lune run test` locally before merging