Setup test runner with Open Cloud (#265)
# Problem

Unit tests can only be run locally with run-in-roblox workflows. We need
to be able to run tests in CI so that PRs can block on failures

# Solution

Open Cloud Luau Execution is the solution! And with the recent upgrade
to Jest 3.10.0 we can finally run tests with loadstring.

We have one longstanding test failure that I don't intend to get working
in this PR, so for now the test job will fail in our checks but is not
required to pass before merging