Use rbxasset for deploying to the Creator Store (#323)
# Problem

We currently handle deployments to the Creator Store with some adhoc
scripts I wrote. We should instead swap these out for more
production-ready code (that I also wrote) in the form of
[rbxasset](https://github.com/Roblox/rbxasset). It's a Luau package that
makes it easier to interact with OC APIs to handle deployments

# Solution

I swapped out the internals of publish-plugin with rbxasset. I've
confirmed that I'm able to deploy to the nightly plugin asset, so we
should be good to go!