Include build info in plugin name (#327)
# Problem

When you have a production version of Flipbook installed it can be hard
to know which Flipbook is which when also developing the plugin.

# Solution

Disambiguate the names. When building for development, Flipbook will
have its name in the Plugins tab updated so you can know which one to
work with