Install packages from Roblox Studio (#329)
# Problem

We want to start using Roblox's internal packages like Foundation and
Signals, but these aren't accessible outside of the Roblox Studio
installation

# Solution

I wrote
[roblox-packages](https://github.com/flipbook-labs/roblox-packages) to
make it possible to extract the Packages folder from the Roblox Studio
installation.

And since we don't want to commit this directly to the repo I've also
setup CI to download Roblox Studio itself to extract
the current package versions

This comes with the very fun caveat that we can't pin versions at all.
So any changes shipped with Roblox Studio will be implicitly adopted by
Flipbook. We're willing to accept the possibility that we'll get random
breaking changes out of nowhere. I certainly don't recommend this for
general use cases though