Pin Roblox packages (#333)
# Problem

It was brought up here
https://github.com/flipbook-labs/flipbook/pull/329#issuecomment-3275078043
that there are actually endpoints that can download historic copies of
Roblox's packages, which means it's possible to pin a specific version
to install.

# Solution

I updated `roblox-packages` to make it possible to install a specific
version of Roblox's packages. Then in Flipbook I've updated rokit.toml
to use the new version, and setup pinning. We'll now consistently be
pulling in packages from `0.690.0.6900721`