Bundle packages with the build files (#337)
# Problem

For #290 I need to get an accurate snapshot of the entire plugin in
plain files. Our rbxm gets built with all of the packages, but I can't
work with the rbxm in this case

# Solution

I setup the build system so that Rojo isn't doing any stitching
together. The final build is output to `build/plugin` including a copy
of all of the packages that are required