Local storage with efficient(ish) disk read/writes (#315)
# Problem

I observed some huge chunks of microprofiler time spent from
GetSetting/SetSetting calls. Presumably each one is making a disk
read/write. And while not in this PR, a change in #290 introduces
frequent read/writes that we need to accommodate

# Solution

Plugin settings are now read from disk only once. In-memory edits are
made from there.

A write is performed on each change, but if needed this can be revised
to batch updates on a timer