Unknown stories (#365)
# Problem

Flipbook would be more accessible if a Storybook was not required to get
started. Other storybook plugins work like this

# Solution

Stories that aren't managed by a storybook will now show up in an
"Unknown Stories" folder. An in-memory storybook is used to preview
these stories like normal.

Unknown stories are free to supply packages on themselves to specify the
renderer to use just like normal

Closes #282