Add option to build for Rotriever (#378)
# Problem

We need some extra support to be able to publish flipbook-core to the
Rotriever registry

# Solution

I've changed some of the terminology, `--target` has been renamed to
`--channel` (for choosing between prod/dev), and a _new_ `--target` flag
has been added to choose if we're building for Rotriever or not. This
could be later extended for other package manages. For example it would
be reasonable to have `wally` and `pesde` targets in the future.

Also each workspace member will now generate a `build` folder inside it.
This makes it easier to deploy _just_ flipbook-core to the Rotriever
registry, and is also just nice to be able to inspect a standalone build
for each member