MUS-2369 Add a workflow for syncing upstream changes (#12)
This still needs some real world testing but the workflows as-is seem to
be doing what we need!

This sets things up so that the main `roblox` branch is largely a
workflow runner. The `upstream` branch will periodically be force-pushed
with the changes from `flipbook-labs/flipbook` and then `roblox` will
have its Rotriever version bumped and deployed with a build from
`upstream`