Fix file watcher getting stuck in a loop (#388)
# Problem

When building with the `--watch` flag the build folders are actually
being included in the files to watch. This results in an infinite build
loop where after the first change, Flipbook will be built, the watcher
will think there's a new change, and then rebuild forever

# Solution

Patched this by adding some exclusion patterns. Also revised how we
discover the files to watch so it should be performing less iterations
over the filesystem

Closes #387