Right-click context menu for Storybooks (#370)
# Problem

We want a right-click context menu for storybooks to allow users to
perform some extra actions

# Solution

I've introduced a new right-click menu for storybooks that includes
Pin/Unpin and "Select in Explorer" options. The former will pin
storybooks to the top of the UI to make it easier to return to
storybooks the user commonly uses. The latter is just a nicety to jump
to the DataModel from Flipbook

Closes #305, closes #291 


https://github.com/user-attachments/assets/2843f78d-6484-4d53-893e-f79a50d11afc