Anonymized usage metrics (#392)
# Problem

We want to start collecting anonymized usage metrics to better
understand things like how many daily active users the plugin has,
session duration, and which features are being used

# Solution

To accomplish this, Flipbook now has a backend service it will send
events to for users that don't opt-out.

The events we send are:
* AppOpened
* AppClosed
* NodePinned
* NodeUnpinned
* PageChanged
* StoryOpened
* StoryClosed
* TelemetryOptedIn
* TelemetryOptedOut

The user is presented with a dialog upon opening the plugin that prompts
them to either remain opted-in or to opt-out.

Afterwards events will be sent to the backend if opted-in.

Closes #356