Whole plugin is broken (#403)
# Problem

It looks like #399 resulted in the entire plugin crashing from a missing
dependency

# Solution

The issue was that ReactReconciler in RobloxPackages was looking for
Promise which had been incorrectly pruned. To fix the issue I've
explicitly included Promise as one of the dependencies to act as a quick
fix.

I'm also installing LuauPolyfill from Wally so we don't need to dip into
RobloxPackages for anything but the essentials