Brand new tree view with Signals (#421)
# Problem

Our tree view in the sidebar has some issues where state updates impact
the entire tree, leading to some nasty bugs like pinning a storybook
resetting the expanded state for all nodes.

# Solution

I'm taking a heavy-handed approach and refactoring TreeViewContext to
use Signals. The end-goal is to have a tree of nodes that can reactively
update themselves in-place, removing the class of issues around state
resetting for the entire tree.

Here it is in action:


https://github.com/user-attachments/assets/67a88bb7-c7a7-4241-bd30-84d291267621

Closes #391, closes #396

Checklist
- [x] Prev selection doesn't clear when selecting another node
- [x] Add filtering (search)
- [x] Pinning correctly adds the node to `Pinned` but also hoists up a
duplicate of that node right below `Pinned`
- [x] Unpinning doesn't update immediately
- [x] Unavailable storybooks don't open
- [x] Last opened story does not get expanded to in the sidebar
- [x] Select, deselect, and reselect the same story
- [x] Unknown stories don't open
- [x] Last opened story does not re-open when closing and reopening the
plugin in the same session