Upgrade Storyteller to 1.4.0 (#435)
# Problem

Storyteller has received a bunch of updates to how it handles the
discovery of Instances in the DataModel. The behavior around
Story/Storybook modules being added/removed/changed should work a lot
more consistently now

# Solution

I went off into Storyteller for a while and fixed up a lot of the
DataModel discovery logic, and introduced changes here in Flipbook to
support it. This includes better in-place edits to StorybookTreeView so
we don't rebuild the entire tree.

This fixes a longstanding issue where the sidebar doesn't always update
based on the state of the DataModel. Like when adding/removing/changing
Storybooks or Stories.

Closes #278 

And the primary reason I did all this was to resolve the release blocker
around orphaned stories not showing up.

Closes #434