Fix/improve code block copying qol (#453)
# Problem

When trying to copy code from our `<CodeBlock />` component, it was full
of the rich text from our Highlighter code.

# Solution

Swap between the original source text and the highlighted version
depending on if it's focused or not. This does still have it's own
little issue where you need to tap on the code block once so it swaps
first before it lets you start highlighting it all.