Flipbook Batteries (#457)
# Problem

We have a lot of useful functions for our Lute scripts that would be
nice to share between our repos

# Solution

I created https://github.com/flipbook-labs/flipbook-batteries to act as
a dumping grounds for useful Luau/Lute modules that we want to share
between repos. This PR consumes FlipbookBatteries and swaps out all of
the functions that have been moved over

# Release notes

Created the
[flipbook-batteries](https://github.com/flipbook-labs/flipbook-batteries)
package which holds all of Flipbook's reusable Luau/Lute modules and is
shared across the other repos we maintain, like Storyteller.