Fix plugin deployments (#464)
# Problem

Deployments for the development plugin have been failing since Dec 22.
Prod will also hit this same issue on the next release

# Solution

When I changed the behavior of the build script to require the
`--plugin` flag to export an rbxm, this inadvertently broke the release
workflow for the prod and dev deployments. I was simply able to update
the publish-plugin script to take in the new flag and verified the dev
plugin deploys

# Release notes

Fixed a bug where the development and production plugins were failing to
deploy