Remove the rest of the Tailwind-based theme (#463)
# Problem

We're now primarily using Foundation which comes with its own themeing.
As such, we're now at a point where we can say goodbye to the
Tailwind-based theme we've been using

# Solution

I've gone ahead and removed the `themes.luau` and `tailwind.luau` files
and adjusted the remaining usage to use Foundation where possible.

We also have an accent color and a GitHub branding color to keep around.
The former I've mapped to one of the purple colors in the Extended
palette, and the latter is just using the same #333333 color as before.
Only reason this is a hook instead of stored in constants is because of
the tokens usage